Extract Reconciler Job Into an own class
Currently the BackgroundThread is an inner class of the
AbstractReconciler and is tightly coupled with it what makes it quite
hard to understand the interaction between both and to properly update a
reconciler on document changes or refactor the code.

This is a first attempt to decouple the things by extracting the inner
class in an own (package private) type to allow further refactoring
operations and proper encapsulation.
